+++ pyproj/_crs.pyx
@@ -6,7 +6,7 @@ from pyproj.compat import cstrencode, pystrdecode
 from pyproj._datadir cimport get_pyproj_context
 from pyproj.enums import WktVersion, ProjVersion
 from pyproj.exceptions import CRSError
-from pyproj.warnings import PyProjDeprecationWarning
+from pyproj.warn import PyProjDeprecationWarning
 
 
 cdef cstrdecode(const char *instring):
